Compile filevector using the -std=c++11 option. This hopefully fixes the comments in this mail by Prof. Ripley: http://lists.r-forge.r-project.org/pipermail/genabel-devel/2013-November/000766.html
Makefile: add -std=c++11 to CFLAGS

fvlib/CastUtils.cpp: Fix (what I think is) the error that aborts the
compilation mention in the e-mail. The error message after compiling
with the updated Makefile is listed at the end of this
message. Basically it complains that there is an ambiguous definition
of the bool isnan(double) function. The <cmath> header puts its
definitions in the std namespace, so the fix I applied simply
explicitely calls the isnan() function from std. 

fvfutil/convert.cpp: adding -Wall to the Makefile gave an error about
comparing an int and an unsigned int. To fix this I decided to cast
the file size (st_size) to an unsigned int. I think that's safe to
do. 

tests/WriteVariablesSpeed.cpp: another error because of the -Wall
option. Again comparing int and unsigned int. Here the fix is
trivial. 



Compile error (mentioned above):

g++ -I fvlib -I fvfutil -g -Wall -std=c++11   -lcppunit fvlib/*.cpp fvfutil/text2fvf.cpp fvfutil/convert.cpp -o bin/convert
fvlib/CastUtils.cpp: In function ‘bool checkNan(void*, int)’:
fvlib/CastUtils.cpp:189:46: error: call of overloaded ‘isnan(double&)’ is ambiguous
    return isnan( *static_cast<double*>(data) );
                                              ^
fvlib/CastUtils.cpp:189:46: note: candidates are:
In file included from /usr/include/features.h:371:0,
                 from /usr/include/x86_64-linux-gnu/c++/4.8/bits/os_defines.h:39,
                 from /usr/include/x86_64-linux-gnu/c++/4.8/bits/c++config.h:426,
                 from /usr/include/c++/4.8/bits/stl_algobase.h:59,
                 from /usr/include/c++/4.8/bits/stl_tree.h:61,
                 from /usr/include/c++/4.8/map:60,
                 from fvlib/CastUtils.cpp:1:
/usr/include/x86_64-linux-gnu/bits/mathcalls.h:235:1: note: int isnan(double)
 __MATHDECL_1 (int,isnan,, (_Mdouble_ __value)) __attribute__ ((__const__));
 ^
In file included from fvlib/CastUtils.cpp:4:0:
/usr/include/c++/4.8/cmath:626:3: note: constexpr bool std::isnan(long double)
   isnan(long double __x)
   ^
/usr/include/c++/4.8/cmath:622:3: note: constexpr bool std::isnan(double)
   isnan(double __x)
   ^
/usr/include/c++/4.8/cmath:618:3: note: constexpr bool std::isnan(float)
   isnan(float __x)
   ^
fvfutil/convert.cpp: In function ‘int main(int, char**)’:
fvfutil/convert.cpp:36:43: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  if (filestatus.st_size < sizeof(data_type))
                                           ^
make: *** [bin/convert] Error 1



git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/filevector@1361 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ fvfutil/convert.cpp
@@ -33,7 +33,7 @@ int main(int argc, char * argv[])
 	stat( filename.c_str() , &filestatus);
 
 
-	if (filestatus.st_size < sizeof(data_type))
+	if ( (unsigned int) filestatus.st_size < sizeof(data_type) )
 		errorLog << "File " << filename <<" is too short to contain an FVF-object." << endl << errorExit;
 
 	fstream dataFile;
